Arreglos en funciones del dao para guardar la informacion del usuario logueado
